BUG: Do not clear text value when default is used
Clearing the text value when the desired value equals the default value
causes an issue when using the control as the value for a parameter in
an ngc call since the SubCallButton uses the text value of the control
as the parameter value. In this scenario there is no way to use the
default value as the settings actual value since the control sets the
text to ''.