BUG: Found shallow copy issue on mdi history
A shallow copy of the framework STATUS mdi value list was being taken which meant changes to it were also happening to the framework managed list. Not what we wanted. This resulted in other changes being needed to support the potentially different list orders based on the mdi history natural-order flag.