Implemented Magic Material system
Like QTPlasmac it is now possible to activate an automatic material definition from a "magic" comment within the gcode.
The "tool" number of 99999 is used to represent this dynamic material definition which is taken from a specifically formatted comment in the ngc file.  Using the same format as the qtplasmac comment with an extra parameter of mt (material thickness).