update jog settings for mill and lathe
this commit updates the jog settings for lathe and mill.  it changes the jog keys for lathe to the arrow keys and inverts the x )up/down) arrow keys based on the LATHE or BACK_TOOL_LATHE ini setting.  it also ensures the keyboard jog safety works and adds the shift key as a rapids override key for fast motion.  I have tested and all seems to work but i need some testers to make sure nothing else broke!